Move cors middleware logic into detectEnviromentForCors fuction
