+++ sonar-project.properties
@@ -16,5 +16,5 @@ sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 
 sonar.test.inclusions=src/**/*.test.ts
 
-sonar.coverage.exclusions=src/server/startServer.ts,src/index.ts,src/database/connectToDatabase.ts
+sonar.coverage.exclusions=src/server/startServer.ts,src/index.ts,src/database/connectToDatabase.ts,src/server/detectEnviromentForCors.ts
